add problem thumbnail and adjust post image logic

+++ backend/src/main/java/com/hoctuan/codingforum/service/post/impl/PostServiceImpl.java
@@ -8,6 +8,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.hoctuan.codingforum.common.BaseServiceImpl;
 import com.hoctuan.codingforum.exception.CustomException;
 import com.hoctuan.codingforum.exception.NotFoundException;
+import com.hoctuan.codingforum.model.dto.post.PostImageDTO;
 import com.hoctuan.codingforum.model.dto.post.PostRequestDTO;
 import com.hoctuan.codingforum.model.dto.post.PostResponseDTO;
 import com.hoctuan.codingforum.model.dto.post.TopicRequestDTO;
@@ -24,6 +25,7 @@ import com.hoctuan.codingforum.service.post.PostService;
 
 import java.util.List;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 @Service
 public class PostServiceImpl extends BaseServiceImpl<
@@ -57,6 +59,9 @@ public class PostServiceImpl extends BaseServiceImpl<
             if(existedPost.getUser().getId() != user.getId()) {
                 throw new CustomException("Yêu cầu không hợp lệ", HttpStatus.BAD_REQUEST.value());
             }
+            if(dto.getPostImage().isEmpty()) {
+                dto.setPostImage(postMapper.toImageDTOs(existedPost.getPostImage()));
+            }
             dto.setTopic(TopicRequestDTO.builder().id(existedPost.getTopic().getId()).build());
         }
 
@@ -74,13 +79,15 @@ public class PostServiceImpl extends BaseServiceImpl<
 
         Post savedPost = postRepository.save(post);
 
-        // delete old images
-        postImageRepository.deleteAllByPost(savedPost);
+        if(!dto.getPostImage().isEmpty()) {
+            // delete old images
+            postImageRepository.deleteAllByPost(savedPost);
 
-        // save new images
-        for(PostImage image : post.getPostImage()) {
-            image.setPost(savedPost);
-            postImageRepository.save(image);
+            // save new images
+            for(PostImage image : post.getPostImage()) {
+                image.setPost(savedPost);
+                postImageRepository.save(image);
+            }
         }
         
         return postMapper.toDTO(savedPost);
